+++ src/utils/index.js
@@ -1,13 +1,25 @@
 'use strict';
+import _ from 'lodash';
 import _clone from 'lodash/clone';
 import _get from 'lodash/get';
-import _isString from 'lodash/isString';
 import _round from 'lodash/round';
 import _pad from 'lodash/pad';
 import _chunk from 'lodash/chunk';
 import _isNaN from 'lodash/isNaN';
+import _has from 'lodash/has';
+import _last from 'lodash/last';
+import _isBoolean from 'lodash/isBoolean';
+import _isString from 'lodash/isString';
+import _isNil from 'lodash/isNil';
+import _isObject from 'lodash/isObject';
+import _isArray from 'lodash/isArray';
+import _isPlainObject from 'lodash/isPlainObject';
+import _forOwn from 'lodash/forOwn';
 import compile from 'lodash/template';
 import jsonLogic from 'json-logic-js';
+
+// Configure JsonLogic
+jsonLogic.add_operation('_', _);
 
 const FormioUtils = {
   jsonLogic, // Share
@@ -18,11 +30,11 @@ const FormioUtils = {
    * @param value
    * @return {boolean}
    */
-  boolValue: function(value) {
-    if (typeof value === 'boolean') {
+  boolValue(value) {
+    if (_isBoolean(value)) {
       return value;
     }
-    else if (typeof value === 'string') {
+    else if (_isString(value)) {
       return (value.toLowerCase() === 'true');
     }
     else {
@@ -35,7 +47,7 @@ const FormioUtils = {
    * @param text
    * @return {Array|{index: number, input: string}|Boolean|*}
    */
-  isMongoId: function(text) {
+  isMongoId(text) {
     return text.toString().match(/^[0-9a-fA-F]{24}$/);
   },
 
@@ -48,12 +60,12 @@ const FormioUtils = {
    * @returns {Boolean}
    *   Whether or not the component is a layout component.
    */
-  isLayoutComponent: function isLayoutComponent(component) {
-    return (
+  isLayoutComponent(component) {
+    return Boolean(
       (component.columns && Array.isArray(component.columns)) ||
       (component.rows && Array.isArray(component.rows)) ||
       (component.components && Array.isArray(component.components))
-    ) ? true : false;
+    );
   },
 
   /**
@@ -70,15 +82,15 @@ const FormioUtils = {
    * @param {Object} parent
    *   The parent object.
    */
-  eachComponent: function eachComponent(components, fn, includeAll, path, parent) {
+  eachComponent(components, fn, includeAll, path, parent) {
     if (!components) return;
     path = path || '';
-    components.forEach(function(component) {
-      var hasColumns = component.columns && Array.isArray(component.columns);
-      var hasRows = component.rows && Array.isArray(component.rows);
-      var hasComps = component.components && Array.isArray(component.components);
-      var noRecurse = false;
-      var newPath = component.key ? (path ? (path + '.' + component.key) : component.key) : '';
+    components.forEach((component) => {
+      const hasColumns = component.columns && Array.isArray(component.columns);
+      const hasRows = component.rows && Array.isArray(component.rows);
+      const hasComps = component.components && Array.isArray(component.components);
+      let noRecurse = false;
+      const newPath = component.key ? (path ? (`${path}.${component.key}`) : component.key) : '';
 
       // Keep track of parent references.
       if (parent) {
@@ -94,13 +106,11 @@ const FormioUtils = {
         noRecurse = fn(component, newPath);
       }
 
-      var subPath = function() {
+      const subPath = () => {
         if (
           component.key &&
           (
-            (component.type === 'datagrid') ||
-            (component.type === 'container') ||
-            (component.type === 'editgrid') ||
+            [ 'datagrid', 'container', 'editgrid' ].includes(component.type) ||
             component.tree
           )
         ) {
@@ -110,28 +120,24 @@ const FormioUtils = {
           component.key &&
           component.type === 'form'
         ) {
-          return newPath + '.data'
+          return `${newPath}.data`
         }
         return path;
       };
 
       if (!noRecurse) {
         if (hasColumns) {
-          component.columns.forEach(function(column) {
-            eachComponent(column.components, fn, includeAll, subPath(), parent ? component : null);
-          });
+          component.columns.forEach((column) =>
+            FormioUtils.eachComponent(column.components, fn, includeAll, subPath(), parent ? component : null));
         }
 
         else if (hasRows) {
-          component.rows.forEach(function(row) {
-            row.forEach((column) => {
-              eachComponent(column.components, fn, includeAll, subPath(), parent ? component : null);
-            });
-          });
+          component.rows.forEach((row) => row.forEach((column) =>
+            FormioUtils.eachComponent(column.components, fn, includeAll, subPath(), parent ? component : null)));
         }
 
         else if (hasComps) {
-          eachComponent(component.components, fn, includeAll, subPath(), parent ? component : null);
+          FormioUtils.eachComponent(component.components, fn, includeAll, subPath(), parent ? component : null);
         }
       }
     });
@@ -144,20 +150,18 @@ const FormioUtils = {
    * @param query
    * @return {boolean}
    */
-  matchComponent: function(component, query) {
-    if (typeof query === 'string') {
+  matchComponent(component, query) {
+    if (_isString(query)) {
       return component.key === query;
     }
     else {
-      var matches = false;
-      for (var search in query) {
-        if (query.hasOwnProperty(search)) {
-          matches = (_get(component, search) === query[search]);
-          if (!matches) {
-            break;
-          }
+      let matches = false;
+      _forOwn(query, (value, key) => {
+        matches = (_get(component, key) === value);
+        if (!matches) {
+          return false;
         }
-      }
+      });
       return matches;
     }
   },
@@ -173,9 +177,9 @@ const FormioUtils = {
    * @returns {Object}
    *   The component that matches the given key, or undefined if not found.
    */
-  getComponent: function getComponent(components, key) {
-    var result;
-    FormioUtils.eachComponent(components, function(component, path) {
+  getComponent(components, key) {
+    let result;
+    FormioUtils.eachComponent(components, (component, path) => {
       if (FormioUtils.matchComponent(component, key)) {
         component.path = path;
         result = component;
@@ -192,9 +196,9 @@ const FormioUtils = {
    * @param query
    * @return {*}
    */
-  findComponents: function findComponents(components, query) {
-    var results = [];
-    FormioUtils.eachComponent(components, function(component, path) {
+  findComponents(components, query) {
+    let results = [];
+    FormioUtils.eachComponent(components, (component, path) => {
       if (FormioUtils.matchComponent(component, query)) {
         component.path = path;
         results.push(component);
@@ -214,9 +218,9 @@ const FormioUtils = {
    * @returns {Object}
    *   The flattened components map.
    */
-  flattenComponents: function flattenComponents(components, includeAll) {
-    var flattened = {};
-    FormioUtils.eachComponent(components, function(component, path) {
+  flattenComponents(components, includeAll) {
+    let flattened = {};
+    FormioUtils.eachComponent(components, (component, path) => {
       flattened[path] = component;
     }, includeAll);
     return flattened;
@@ -229,12 +233,12 @@ const FormioUtils = {
    *
    * @returns {boolean} - TRUE - This component has a conditional, FALSE - No conditional provided.
    */
-  hasCondition: function(component) {
-    return (
-      (component.hasOwnProperty('customConditional') && component.customConditional) ||
-      (component.hasOwnProperty('conditional') && component.conditional && component.conditional.when) ||
-      (component.hasOwnProperty('conditional') && component.conditional && component.conditional.json)
-    ) ? true : false;
+  hasCondition(component) {
+    return Boolean(
+      (component.customConditional) ||
+      (component.conditional && component.conditional.when) ||
+      (component.conditional && component.conditional.json)
+    );
   },
 
   /**
@@ -246,7 +250,7 @@ const FormioUtils = {
    * @returns {Number}
    *   Parsed value.
    */
-  parseFloat: function(value) {
+  parseFloat(value) {
     return parseFloat(_isString(value)
       ? value.replace(/[^\de.+-]/gi, '')
       : value);
@@ -261,7 +265,7 @@ const FormioUtils = {
    * @returns {String}
    *   Value formatted for Currency component.
    */
-  formatAsCurrency: function(value) {
+  formatAsCurrency(value) {
     const parsedValue = this.parseFloat(value);
 
     if (_isNaN(parsedValue)) {
@@ -273,11 +277,9 @@ const FormioUtils = {
       .split('.');
     parts[0] = _chunk(Array.from(parts[0]).reverse(), 3)
       .reverse()
-      .map(function(part) {
-        return part
-          .reverse()
-          .join('');
-      })
+      .map((part) => part
+        .reverse()
+        .join(''))
       .join(',');
     parts[1] = _pad(parts[1], 2, '0');
     return parts.join('.');
@@ -291,7 +293,7 @@ const FormioUtils = {
    * @returns {string}
    *   String with escaped RegEx characters.
    */
-  escapeRegExCharacters: function(value) {
+  escapeRegExCharacters(value) {
     return value.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, "\\$&");
   },
 
@@ -305,27 +307,27 @@ const FormioUtils = {
    * @param data
    *   The full submission data.
    */
-  checkCalculated: function(component, submission, data) {
+  checkCalculated(component, submission, data) {
     // Process calculated value stuff if present.
     if (component.calculateValue) {
-      if (typeof component.calculateValue === 'string') {
+      if (_isString(component.calculateValue)) {
         try {
           const util = this;
-          data[component.key] = eval('(function(data, util) { var value = [];' + component.calculateValue.toString() + '; return value; })(data, util)');
+          data[component.key] = eval(`(function(data, util) { var value = [];'${component.calculateValue.toString()}; return value; })(data, util)`);
         }
         catch (e) {
-          console.warn('An error occurred calculating a value for ' + component.key, e);
+          console.warn(`An error occurred calculating a value for ${component.key}`, e);
         }
       }
       else {
         try {
-          data[component.key] = FormioUtils.jsonLogic.apply(component.calculateValue, {
+          data[component.key] = this.jsonLogic.apply(component.calculateValue, {
             data: submission ? submission.data : data,
             row: data
           });
         }
         catch (e) {
-          console.warn('An error occurred calculating a value for ' + component.key, e);
+          console.warn(`An error occurred calculating a value for ${component.key}`, e);
         }
       }
     }
@@ -343,48 +345,46 @@ const FormioUtils = {
    *
    * @returns {boolean}
    */
-  checkCondition: function(component, row, data) {
-    if (component.hasOwnProperty('customConditional') && component.customConditional) {
+  checkCondition(component, row, data) {
+    if (component.customConditional) {
       try {
-        var script = '(function() { var show = true;';
-        script += component.customConditional.toString();
-        script += '; return show; })()';
-        var result = eval(script);
+        const script = `(function() { var show = true;${component.customConditional.toString()}; return show; })()`;
+        const result = eval(script);
         return result.toString() === 'true';
       }
       catch (e) {
-        console.warn('An error occurred in a custom conditional statement for component ' + component.key, e);
+        console.warn(`An error occurred in a custom conditional statement for component ${component.key}`, e);
         return true;
       }
     }
-    else if (component.hasOwnProperty('conditional') && component.conditional && component.conditional.when) {
-      var cond = component.conditional;
-      var value = null;
+    else if (component.conditional && component.conditional.when) {
+      const cond = component.conditional;
+      let value = null;
       if (row) {
         value = this.getValue({data: row}, cond.when);
       }
-      if (data && (value === null || typeof value === 'undefined')) {
+      if (data && _isNil(value)) {
         value = this.getValue({data: data}, cond.when);
       }
       // FOR-400 - Fix issue where falsey values were being evaluated as show=true
-      if (value === null || typeof value === 'undefined') {
+      if (_isNil(value)) {
         return false;
       }
       // Special check for selectboxes component.
-      if (typeof value === 'object' && value.hasOwnProperty(cond.eq)) {
+      if (_isObject(value) && _has(value, cond.eq)) {
         return value[cond.eq].toString() === cond.show.toString();
       }
       // FOR-179 - Check for multiple values.
-      if (value instanceof Array && value.includes(cond.eq)) {
+      if (_isArray(value) && value.includes(cond.eq)) {
         return (cond.show.toString() === 'true');
       }
 
       return (value.toString() === cond.eq.toString()) === (cond.show.toString() === 'true');
     }
-    else if (component.hasOwnProperty('conditional') && component.conditional && component.conditional.json) {
+    else if (component.conditional && component.conditional.json) {
       return jsonLogic.apply(component.conditional.json, {
-        data: data,
-        row: row
+        data,
+        row
       });
     }
 
@@ -400,36 +400,31 @@ const FormioUtils = {
    * @param {String} key
    *   A for components API key to search for.
    */
-  getValue: function getValue(submission, key) {
-    var data = submission.data || {};
-
-    var search = function search(data) {
-      var i;
-      var value;
-
-      if (!data) {
-        return null;
-      }
-
-      if (typeof data === 'object' && !(data instanceof Array) && data !== null) {
-        if (data.hasOwnProperty(key)) {
+  getValue(submission, key) {
+    const search = (data) => {
+      if (_isPlainObject(data)) {
+        if (_has(data, key)) {
           return data[key];
         }
 
-        var keys = Object.keys(data);
-        for (i = 0; i < keys.length; i++) {
-          if (typeof data[keys[i]] === 'object') {
-            value = search(data[keys[i]]);
-          }
+        let value = null;
 
-          if (value) {
-            return value;
+        _forOwn(data, (prop) => {
+          const result = search(prop);
+          if (!_isNil(result)) {
+            value = result;
+            return false;
           }
-        }
+        });
+
+        return value;
+      }
+      else {
+        return null
       }
     };
 
-    return search(data);
+    return search(submission.data);
   },
 
   /**
@@ -439,7 +434,7 @@ const FormioUtils = {
    * @param data
    * @returns {XML|string|*|void}
    */
-  interpolate: function(string, data) {
+  interpolate(string, data) {
     const templateSettings = {
       evaluate: /\{\%(.+?)\%\}/g,
       interpolate: /\{\{(.+?)\}\}/g,
@@ -458,19 +453,21 @@ const FormioUtils = {
    * @param name
    * @returns {string}
    */
-  uniqueName: function(name) {
-    var parts = name.toLowerCase().replace(/[^0-9a-z\.]/g, '').split('.');
-    var fileName = parts[0];
-    var ext = '';
-    if (parts.length > 1) {
-      ext = '.' + parts[(parts.length - 1)];
-    }
-    return fileName.substr(0, 10) + '-' + this.guid() + ext;
+  uniqueName(name) {
+    const parts = name.toLowerCase().replace(/[^0-9a-z\.]/g, '').split('.');
+    const fileName = parts[0];
+    const ext = parts.length > 1
+      ? `.${_last(parts)}`
+      : '';
+    return `${fileName.substr(0, 10)}-${this.guid()}${ext}`;
   },
 
-  guid: function() {
-    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-      var r = Math.random()*16|0, v = c === 'x' ? r : (r&0x3|0x8);
+  guid() {
+    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+      const r = Math.random()*16|0;
+      const v = c === 'x'
+        ? r
+        : (r&0x3|0x8);
       return v.toString(16);
     });
   }